+++ src/pages/SignIn/index.tsx
@@ -1,9 +1,9 @@
 import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import RegisterForm from "../../components/auth/RegisterForm/RegisterForm";
+import LoginForm from "../../components/auth/LoginForm";
 import { S } from "./style";
 
-export default function SignUp() {
+export default function SignIn() {
   const navigate = useNavigate();
 
   useEffect(() => {
@@ -15,10 +15,10 @@ export default function SignUp() {
 
   return (
     <main>
-      <h1 className="ir-hidden">회원가입 페이지</h1>
+      <h1 className="ir-hidden">로그인 페이지</h1>
       <S.Container>
-        <S.Title>회원가입</S.Title>
-        <RegisterForm />
+        <S.Title>로그인</S.Title>
+        <LoginForm />
       </S.Container>
     </main>
   );
